requirements.txt
spotipy==2.23.0
pyncm==1.6.8.9.1
PyYAML==6.0.2
pybase64==1.2.3
tqdm==4.66.3
Unidecode==1.3.8
loguru